Fixed inconsistent parameters
Unfortunately, at the moment, my design is over-determined. The
ADDR_WIDTH, BYTE_ADDR_WIDTH, and DATA_WIDTH parameters have to
follow a certain relation, and if they don't, it breaks the
logic.

It's a simple fix, once I figure out how to do log base 2 in
Verilog

One more thing: I already knew this, but I've forgotten about this
detail in the past. The forwarder cannot say it is done on the same
cycle as its last rd_en. You have to wait for the memory to appear;
if you say you're done, then the buffer is immediately wired to
another core and you won't get your data